Implemented Default Classes for Grid Interface. >From now on all implementations of the Grid interface should be derived from the Default Classes, i.e. GridDefault. All Default classes implement some extra functionality using the slim interface described in the interface classes. For example the method subIndex of the class EntityDefault<codim=0...> returns the global index of a sub entity with codim=cc and a local number. The Implenentation is done via the interface method entity of the Entity Interface.
The class GridDefault provides an LeafIterator. This iterator uses the
LevelIterator and Hierarchical Iterator to iterate over all leaf elements
until a given maxlevel.
The LeafIterator has the same Interface as the LevelIterator.

Furtermore added some methods for writing the grid to an file. But im still
working on it. Therefore explanations later.

[[Imported from SVN: r88]]