This patch brings you second order isoparametric elements.
Not everything is working yet, but the complete infrastructure
is there and you can already do simple finite elements on
simplex grids.  Methods which are not working properly yet
will throw a NotImplemented-exception instead of tacitly
returning wrong results.  They will be completed one-by-one.

The performance penalty seems to be minimal, except for the
methods of the intersection geometries.  There is a way to
get those methods back to speed, but it involves some refactoring
and I will do that separately.

[[Imported from SVN: r3774]]