Implemented Dune::Intersection.
The Dune::IntersectionIteratorDefault offers an implementation of
dereference, so that the existing implementations for
IntersectionIterator can also be used for both IntersectionIterator
and Intersection.

The Intersection can parametrized with what was the
IntersectionIteratorImp. This allows an easy transition. The Grid
Traits now also export the types LeafIntersection and
LevelIntersection and take two new parameters for the
LeafIntersectionImp and LevelIntersectionImp. I updated all grids
using the Dune::GridTraits class, Alberta and Alugrid use their own
Traits. I didn't want to break anything. This means that someone from
Freiburg has to add these typedefs to the Traits.

[[Imported from SVN: r4077]]