This patch adds a second constructor to the GridFactory. The new constructor accepts a non-const pointer to a grid object. This object should then be filled with input by the factory.
The reason for this is that some grids have more than
one constructor.  One example is UGGrid, where you can construct
grids using the default constructor, but you can also provide
one integer, which is the amount of memory UG uses internally.
(purist will say that this is a flaw in UG.  I agree, but it
can't be helped with a reasonable amount of work).

If more than one constructor are present the factory doesn't
know which one to call, and the user wants to keep the
possibility to use all constructors.  Therefore the factory
class now accepts pre-constructed but empty objects.


[[Imported from SVN: r4136]]