remove the use of ALBERTA traverse stacks in DUNE bindings
ALBERTA's traverse stacks were manipulated in DUNE directly. In ALBERTA 2.1
this is no longer possible, because they are a private stucture. On the other
hand, the hierarchic iterator always had to copy to preserve the stack of the
originating entity.

All this is no longer needed. The (wrapped) element info structures now point
to their father structure. Reference counting is used to ensure that an element
info structure is not freed as long as it might still be required.

[[Imported from SVN: r4583]]