Bugfix: properly clean up UG data structures before creating a new grid object
This fixes 75, but it needs a small patch to UG as well.  This patch is in
our UG repository, however I can't properly distribute it as long as the
Wittum UG is unavailable.  Therefore one crucial line of this patch is
commented out, the bug is not actually patched, but almost...

sigh...


[[Imported from SVN: r4898]]