make interval block parse all intervals at once and make sure there is only one if YaspGrid is to be used
It is possible to have multiple intervals in the interval block. If you did this
with YaspGrid, you only got the first one (without a word of warning). Now, the
DGFParser will in this case and tell you what's wrong.

Also, the vertex and cube generation methods have been extended to work in arbitrary
space dimension (produces exactly the same result for dim=1,2,3)

[[Imported from SVN: r4916]]