reduce preprocessor magic for ALBERTA's world dimension
ALBERTA's world dimension is now selected exclusively by the preprocessor macro
ALBERTA_DIM. In ALBERTA_CPPFLAGS, this is set to the content of the environment
variable ALBERTA_DIM. If DUNE is configured with --with-grid-dim or --with-world-dim,
the variable ALBERTA_DIM defaults to $(GRIDDIM).

This way we don't break the GRIDDIM / WORLDDIM magic, but it becomes transparent
how ALBERTA's world dimension is controlled.

[[Imported from SVN: r5006]]