Use the type of thr grid itself (AlbertaGrid<dim,dimworld>) as dummy template parameter to CollectiveCommunication
Currently, "int" is used as dummy parameter.  This is problematic
because there is a specialization of CollectiveCommunication for the
template parameter MPI_Comm, which is identical to int in some MPI
implementations (for example, MPICH).  This specialization is used if
we use AlbertaGrid alongside with a parallel grid, but the
specialization does not have a default constructor, so compilation
will fail.

It would be nice if this patch could be merged to version 1.2 for a
potential future bugfix release.


[[Imported from SVN: r5077]]