1) fix refineOptions 2) Implement mark/adapt stategy for globally refined grid
mark(refCount) marks the entity for adaptation and requires all other
entities to be refined as well. If refCount is smaller than the
refCount of any other entity (read the refCount of any previous call)
the previous refCount is used.

This would bring the mark/adapt interface of globalrefine meshes as
close as possible to that of adaptive meshes and would allow the
generic use of error estimators.

While this semantic seems to be fine with the docu in grid.hh it fails
in the grid test (gridcheck.cc:962) as a later call to mark does not
result in the same mark-state as the one requested but only in
mark-state >= refCount.

[[Imported from SVN: r5478]]