* disable broken part in the test * test compiles again (also with old interface used) * fix the mappers
OK this is nasty:

we want the mappers to implement the new interface and to use a
default implementation of the old interface.

The ideas is that the Mapper base class offers the default
implementation including renumbering etc. Now the problem is that the
old and the new method have the same name. The implementation of the
new method shadows all inherits map methods, thus we have add a using
statement at the beginning of our mapper implementation.

The same changes will be necessary for all indexsets and idsets.

[[Imported from SVN: r5518]]