BoundarySegments are now handed to the grid factory using a shared_ptr instead of a raw pointer.
Currently this concerns UGGrid, AlbertaGrid and ALUGrid.
For the latter two, I didn't know how to properly adapt the code
beyond the actual interface.  Hence these implementations now
extract the raw pointer from the smart pointer and use that.
This should be fixed eventually.  The corresponding parts of
the code are marked with a \todo.


BTW: I really enjoyed the 

   DUNE_THROW(GridError,"Fuck gmsh");

:-)


[[Imported from SVN: r6251]]