Improve alberta.m4 in the following ways: * add a switch --disable-alberta-libcheck to explicitly disable check for   ALBERTA libs (default = perform check) * build the DUNE AlbertaGrid library exactly for those libraries present in   ALBERTA
When configuring ALBERTA, it is possible to deactivate building of libraries
for dimension 1, 2, or 3 individually.
Moreover, in ALBERTA 3.0 it is possible to additionally build libraries for
dimensions 4 to 9 (configure switch --enable-dim-of-world="DIM1 ... DIMk").
Note: This does not influence the maximal grid dimension (which is 3).

alberta.m4 now checks whether libalberta_Nd is present. If it is, it adds
DUNE support for this world dimension.

[[Imported from SVN: r6362]]