Allow header files for grid selection to be generated by configure script
The preprocessor magic for selecting the grid currently cannot be extended to
support grids from modules other that dune-grid. Moreover, the code is mostly
'cut & paste', which is a problem for maintenance.

This patch overcomes both difficulties. The files 'gridtype.hh' and
'dgfgridtype.hh' are generated for each module by configure. They contain
the switches for all the grids found. To add an additional grid, just add
a call to DUNE_DEFINE_GRIDTYPE in the main m4-file of the module that
contains it (see e.g. dune-grid/m4/dune_grid.m4 or dune-grid/m4/alberta.m4)
All the preprocessor stuff is then generated from this line.

[[Imported from SVN: r6391]]