Add information about how the boundary segments are numbered.
Since the grid is 1d and connected, this information is a single boolean
'reversedBoundarySegmentNumbering' which stores 'false', if the left
boundary is 0 and the right one is 1, and 'true' if it is the other
way around.

Getting this information to the intersections is the tricky part,
because they don't have access to the grid.  Therefore now each and
every element keeps a copy of the flag.

[[Imported from SVN: r6440]]