Use the GenericGeometries facilities to compute intersection normals.
This may be a bit slower now, but the code is *a lot* shorter.
The current implementation sets up a BasicGeometry for the inside element,
then computes local coordinates with respect to this element and then
uses the normal-computing method of BasicGeometry.  Speed could be
regained if there was something like a generic IntersectionGeometry,
which would save me getting corner coordinates I don't need and
doing the coordinate transformation.

[[Imported from SVN: r6536]]