revert patch 6565
I agree with Oliver, that the code would look nicer. The problem with static
ifs, though, is that both branches have to be instanciated. Conversion< prism, 0 >
then instanciates Conversion< simplex, -1 > and so on.

In my opinion, we should discuss the future of the GeometryType before we start
discussing how the geometry type conversion stuff can be beautified.

Oh, yes, and thanks for extensive testing. The bug could have been found by
configurung with ALBERTA.

[[Imported from SVN: r6566]]