[VTKWriter] Remove the superficial similarities between VertexIterator and CornerIterator and document their differences.
The two iterators looked very much the same superficially with the exception
that CornerIterator always iterates over every corner of every element, while
VertexIterator may skip corners which correspond to a vertex it has already
seen in another element, depending on whether conforming or nonconforming data
mode is in effect.

The is however another very fundamental difference, which was hidden very
deeply: VertexIterator iterates over the corners of an element in the order
given by Dune's reference elements, while CornerIterator iterates in the order
given by VTK's reference elements.

This whole mess was made even more confusing because both iterators contained
an identical method id() which obvious assumed VTK's ordering, while other
methods clearly assumed Dune's ordering.  For the correctness of the code this
did not matter, since in the case of the CornerIterator the methods assuming
Dune-ordering where never used, and in the case of the VertexIterator the
id()-method was nether used.

To reduce similar days-long misunderstandings in the future, I have removed
the unused methods from both iterators and clearly documented the differences.

[[Imported from SVN: r6745]]