change template arguments for ALU3dGrid (should be unnoticable to user)
the template arguments dim and dimworld of ALU3dGrid have been removed (they had
to be 3 anyway) and an extra template argument Comm (modelling the communicator,
either No_Comm or MPI_Comm) has been added. The addition of this template argument
will enable us to move parts of the ALUGrid implementation into the lib, because
the implementation is no longer selected by a preprocessor switch.

For the user this change should go unnoticed, since he uses ALUSimplexGrid or
ALUCubeGrid anyway. Hence, there is no need for any backward compatibility.

Please note: There are still some places where the preprocessor switches the
implementation. This is ok, as long as the default value for the template
argument Comm is used (ALUSimplexGrid and ALUCubeGrid currently do this).

[[Imported from SVN: r6951]]