[StructuredGridFactory] Work in parallel programs.
Introduce a traits class FactoryNeedsInsertOnAllRanks<Grid> with a boolean
member constant 'value'.  This is 'false' per default and 'true' for UGGrid.

Use this traits class and the rank from MPIHelper's communicator to determine
whether to skip insertions to the GridFactory.

The call to loadBalance() is still left to the caller.

[[Imported from SVN: r7078]]