[UGGrid] Make it possible to statically set UG's default heap size.
Make the heap size a static variable of the grid class.  Introduce a static
method setDefaultHeapSize() to change it.  The contructor taking a heap size
argument is now deprecated, and simply sets the default heap size, before
doing the same as the default contructor.

Thanks Oliver for help when coding the patch.

[[Imported from SVN: r7100]]