Transfer the entire coarse grid information from rank 0 to all other processes before grid construction.
At the 2010-11 Dune meeting in Münster the following was decided:
Parallel grids that are set up by providing a coarse grid on rank 0
and then loadbalancing should not require that the grid should be
provided to all processes directly (currently UGGrid does that).
This patch makes the rank 0 process broadcast the entire coarse
grid information from rank 0 to all other processes before
starting the actual grid creation.  That way, the user has to
supply the coarse grid only on rank 0 (as the specification says),
but still the UGGridFactory has all relevant information available
on all processes.


[[Imported from SVN: r7124]]