Add a method 'contains' to the GridView interface.
This has been discussed in FS 515, and decided upon at the
2010 Münster meeting.  The new method is a copy of the previous
method on the IndexSet, both semantically and syntactically.
In particular, there is no version with a PartitionType
template parameter.  I leave that to people who deal with
parallel grids more often.

The GridView interface class currently forwards the call
to 'contains' to the corresponding method of the index set.
This can of course be suboptimal, but cannot be changed
unless all the grid view implementations implement 'contains'
themselves.  I will work towards that goal in the nearer
future.

[[Imported from SVN: r7220]]