First implementation of EntitySeed. Interface class is missing currently.
For all grids not implementing this class yet the default value will be EntityPointer. 
I added two new template parameters to GridTraits, LocalGeometryImp which defaults to GeometryImp,
and EntitySeedImp which defaults to EntityPointerImp and then the default EntitySeed
type is just EntityPointer.  

The EntityDefaultImplementation now contains a method seed which is already deprecated
to indicate that there has to be some work done. Also the method entityPointer on the
GridDefaultImplementation. 

All tests (except UG which was not testet due to missing installation) in dune/grid/test
compiled and worked. Since I did not have to change anything in the grid implementations
itself I'll expect UGGrid to compile.



[[Imported from SVN: r7228]]