Use the UG Tag to differentiate between element types, instead of the number of sides
In the UGGridRenumberer: the number of sides is ambiguous, as both pyramids
and prisms have five of them.  This hasn't been a problem so far, because
the faces of pyramids and prisms didn't get renumbered anyways.  With the
new Dune numbering this changes and the nSides parameter needs to go.

[[Imported from SVN: r7271]]