Introduce a new class UGGridLocalGeometries that implement local geometries.
Previously, local and global geometries were handled by the same class,
namely UGGridGeometry.  This was a bit messy, and the code contained
a dynamic 'if' right in the center.  Splitting off the local geometries
into a new class allows to clean up the code.  In particular it allows
to remove the 'if'.  UGGrid may just get a tiny bit faster.

The new class is simple: it forwards everything to a generic geometry.

[[Imported from SVN: r7657]]