Modify the test for the upcoming 1.3 version of libpsurface.
libpsurface-1.3 will have its entire code contained in a namespace.
The m4 test has to be adjusted for this.  In order not to force
users to upgrade their libpsurface right away, the new test
tests for the old (pre 1.3) and new version.  Then a cpp macro
PSURFACE_NAMESPACE with the correct value is defined.
This procedure is temporary.  In dune-grid 2.2 at the latest,
libpsurface 1.3 or newer will be mandatory, and the namespace
handling will be hardwired into dune-grid.

[[Imported from SVN: r7711]]