Fix a bug in the UGGrid handling of boundary segment indices.
This has been reported by Lars Lubkoll: If you use more than
one UGGrid, then the boundary segment indices of the grids
get corrupted.

The fix is simply, but still the patch is invasive.  To fix
the bug, UG's global currBVP variable needs to be set to the
current grid to obtain the correct indices.  That's one
additional line of code.  However, the pointer to the current
grid is not available in the UGGridIntersections class...

This is what makes the present patch so long: it basically
make the grid known to almost all UGGrid data structures.
This adds a little more overhead, but I really don't see
any other way to fix the bug.

[[Imported from SVN: r7800]]