Remove the test 'testgeo'
I am not convinced of its usefulness.  If somebody thinks it
should be kept that's fine, but it needs to go somewhere else.
Its suitable location depends on its true purpose.  Does it
change a property of a grid?  Then it should go to where the
other grid tests are.  Or does it test the functioning of the
BasicGeometry class?  Then it should go into dune-geometry.

[[Imported from SVN: r8279]]