[dgfparser/utils] Backport patch r8228 from mnolte's branch.
The original commit message was:

reenable the preprocessor magic for dgf2dgf and gmsh2dgf

Though onedgrid is the default now, it is (again) possible to exchange the
grid by writing, e.g.,
  make GRIDTYPE=ALBERTAGRID GRIDDIM=2 WORLDDIM=3
for the mentioned utilities. This was disabled by Christoph's previous patch.

Remark: Both utilities are pretty useless for OneDGrid. It can at most be a
        default, but should not be hardwired.

[[Imported from SVN: r8296]]