Implement the full geometry interface for static cube-cube refinement
The current implementation uses a BasicGeometry for the task.
This means that the code is not as fast as it could be.
I might improve that in the future.  At least all refinement
variants comply with the geometry interface now.

Unlike the other refinement implementations, the geometry handed
out is a GeometryImplementation class directly.  All other
refinements hand out Geometry<GeometryImp>.  I don't know
what is better in this case.  The user won't notice the
difference anyway.

[[Imported from SVN: r8320]]