Decruft: remove support for the UG lgm domain manager
It is a non-standard extension to the grid interface which
was supposed to smooth the way for some transitions from
pure UG to dune+UGGrid.

Very few people know how it works, so maintenance would be
extremely difficult if we were to seriously maintain it.
However it turns out that this code wouldn't even compile
all the way back to at least dune-grid-2.1. I have therefore
strong doubts that anybody is using it at all.

[[Imported from SVN: r8330]]