Rudimentary implementation of dynamic load balancing (part I)
This patch is a first part of a simple implementation of
dynamic load balancing for UGGrid.  In this implementation,
UGGrid sends source processors along with elements and vertices,
and then an all-to-all communication is done to transfer the
actual data.

Needless to say, this is neither very elegant nor does it
scale.  Still, it does work (modulo bugs), and allows us
(i.e. Bernd and me) to understand the problem better.
After looking at the UG internals some more I believe that
true dynamic load balancing can be done for UGGrid (modulo
those never quite precisely stated bugs in UG itself).
But that will take some more time and effort.

[[Imported from SVN: r8497]]