Add a few warnings about a potential issue/bug with ids on parallel UGGrids
The Dune grid interface mandates that two entities on different levels have
the same id if they are copies.  Therefore the UGGrid id/subId methods
look through the ancestry of a given entity to see if there are copies.
However, UG does so-called vertical load-balancing, which means that the
ancestry may be distributed across different processors.  Since the current
code does not take that into account, wrong ids will be returned in the
presence of vertical load-balancing.  This is a potentially severe issue,
because users do not get error messages, only ids that are tacitly wrong.

Fixing this properly is not easy either, though.  Currently I don't really
even know how to produce a test case.

[[Imported from SVN: r8515]]