Merge local and global id sets into a single class and object.
Because as it turns out, in the case of UGGrid it is best to have
both global and local ids be the same.  Global ids are not more
difficult to obtain than local ones.  On the contrary, it is not
clear how local ids should be constructed, and maintained through
load balancing processes.

[[Imported from SVN: r8594]]