Bugfix: take the ddd global id of the vertex instead of of the node
In UG, node ids are different on each level, while there is only a single
vertex (hence with a single id) for all nodes at one position on all
levels.  Since the Dune grid specification says that copies of an entity
on different levels have the same id, we need to take the UG vertex id,
instead of the node id.

Thanks to Bernd Flemisch for debugging this.


[[Imported from SVN: r8614]]