Cleanup: Remove three out of four default template parameters from the GridTraits struct
Rationale: The third of these parameters was EntityPointerImp for EntitySeedImp.
Since the fallback implementation of EntitySeed has just been removed, leaving
the EntityPointerImp default there is plain wrong.  Yes, it doesn't hurt,
but it is fairly measleading.
However, if you remove a default parameter, all the ones in front of it need
to go, too.  These are DefaultLevelGridViewTraits and DefaultLeafGridViewTraits.
There is no strong reason to remove them, but since all grids within my reach
do implement EntitySeed, none of them can use the Default*GridViewTraits parameters.
Hence it is safe to remove the defaults.

[[Imported from SVN: r8760]]