Add new constructors that use std::bitset instead of FieldVector<bool>
Because FieldVector<bool> is ugly.  The bool class does not implement
the methods of a field, hence it is conceptual nonsense to instantiate
a FieldVector of bool.

The same holds for the FieldVector<int> still in the interface,
but that is more difficult to get rid of.

The old constructors are still there for backward-compatibility.

[[Imported from SVN: r8773]]