Bug fix in YaspEntityPointer(YaspEntity) constructor.
When an EntityPointer 'ep' is created from an entity 'other'
the entity 'e' stored in 'ep' was tied to 'other' instead
of the data in 'ep' itself. Hence 'e' and 'ep' where inconsistent
if 'ep' was reassigned.

This fix solves the problem revealed by the new checkFather()
method in r8891.

[[Imported from SVN: r8892]]