[CMAKE] Fixes FindGrape if no X11 or OpenGL is present.
Previously, both X11 and OpenGL were required when searching
for them using find_package. This resulted in a fatal error
during the configure stage if either of them was not found.
As result dune-grid was not usable without X11 and OpenGL.

This patches fixes this situation by making X11 and OpenGL
optional. Thus configure succeeds even without them as it
should.