[AlbertaGrid] remove support for generic geometries again
AlbertaGrid had experimental support for using the generic geometries
(enabled by a preprocessor macro known to few people).
As we decided on the developer meeting in Aachen to remove the old
generic geometry implementation in favor of the new MultiLinearGeometry,
we need to adapt AlbertaGrid. For performance reasons, we keep the
specialized AlbertaGrid implementation and simply drop the support for the
generic geometry code.