[Cmake] Determine UG_DIR based on UG_ROOT and add mpi flags for parallel dune.
With this patch is suffices to specify the root directory of the UG installation
and CMake will guess the directory where the CMake configuration file resides.
In addition the mpi flags for parallel UG are now automatically added.

Note: You need revision 8524 of our UG fork for this to work. Future patch
levels (>10) willl incorporate these changes.