Implement a dedicated LeafIndexSet class, instead of using the max-LevelIndexSet
Previously, in a rage of cleanup, I had removed the class that implements
the LeafIndexSet for YaspGrid.  Instead, the LevelIndexSet for the maxLevel
was used.  However, that creates a problem when a LeafIndexSet is copied,
and then the max level removed from the grid.  In that case, the copied
LeafIndexSet points to a level that is no longer existent.  This lead to
crashes in adaptivefinitevolume (in dune-grid-howto), see FlySpray #1276.

This patch puts the separate LeafIndexSet class back in.