[Grid tests] Provide interface check for jacobian related return types.
The Dune grid interface allows for implementation-defined return types
of the Dune::Geometry member functions:

  JacobianInverseTransposed &jacobianInverseTransposed ( const LocalCoordinate &local );
  JacobianTransposed &jacobianTransposed ( const LocalCoordinate &local );

The return types 'JacobianInverseTransposed', 'JacobianTransposed' are
only required to partially implement the Dune::DenseMatrix interface
(in particular methods for the various matrix-vector operations). In
addition, they must be convertible into a Dune::FielMatrix.

This patch provides additional documentation and an interface check for
the return types of the aforementioned functions.