make test-alberta3d-refine pass again (FS#1363)
Up to 02cc041e660ca569a2b40c4c44d6384c650ffeb3, calling markLongestEdge
on the grid factory was the default behavior of createGrid. This
approach has 2 disadvantages:
(a) Longest-Edge bisection might result in very non-local refinements
(b) In 3d, it is not always possible to mark the second-longest edges as
    refinement edges for level 1 (i.e., if the second longest edge does
    not share a common vertex with the the longest one)
As a result, this algorithm might potentially even break a macro grid
originally compatible with Alberta's recursive bisection and, therefore,
is now optional (the user has to call markLongestEdge).
However, markLongestEdge works perfectly for Kuhn-Triagulations, because
the longest edge is always the diagonal of the cube and the second longest
edges are exactly the face diagonals of the cube.

This patch fixes the failing test by calling markLongestEdge manually,
thus compensating the change in the grid factory.