[bugfix] only write BOUNDARYSEGMENTS block, if experimental grid extensions are enabled
On the last meeting, we decided to make the boundaryId an experimental
grid extension. As a result, writing the BOUNDARYSEGMENTS block without
additional user data to a DGF file no longer makes sense in this case.
As the code for writing this information relies on the method
boundaryId, it did not even compile. This patch resolves the situation
by writing the BOUNDARYSEGMENTS block if and only if experimental grid
extensions are enabled.