Remove the test for FS 940
The test relied on the boundaryId method, which disappeared recently
into the realm of EXPERIMENTAL_GRID_EXTENSIONS.  Unfortunately, nobody
seems to remember what the text actually did, so fixing it means some
effort on the part of somebody.  This may not be worth it, hence I
simply remove the test.