[bugfix] resolve ambiguity in constructor call std::bitset
For gcc-4.5 calling std::bitset(0) is ambigious for the new C++11 standard.
int 0 can be implicitly cast into char* as well as into long long unsigned int.
This is a bug in gcc-4.5 because the constructor from char* should actually be templatized over
the character type.
This patch calls the default ctor of std::bitset, which also fills the bitset with zeroes.