Introduce new methods levelGridView and leafGridView for each grid
There has been much discontent over the fact that the leafView method
is called 'leafView', but the corresponding type is called 'LeafGridView'.
(and similarly for level views).  At the dev meeting in Aachen we
decided to get rid of this inconsistency.  Hence, as a first step,
this patch introduces new methods leafGridView and levelGridView,
which mirror the old methods.  No deprecation has been done in this
patch, but it will follow shortly.

but the correspondin