[bugfix] Fix hierarchic iterator for the refineOptions keep_ovlp==false case
With keep_ovlp==false, not all entities have 2^d sons. The hierarchic iterator
ignored that fact. This was a bug for the last 10 years, but nobody realized
because the resulting entities where geometrically correct (though not in the grid).
In a tensorproduct grid, these were not gemoetrically correct anymore and thus
the geometryInFather test failed with that refinement option.