[bugfix] introduce compatibility code for old constructors
the deprecated constructors pass their arguments to the torus in the
initializer list. having additional deprecated constructors on the
torus is verz messy, so i introduced a compatibility function
convertFieldVectorToArray, which is immediately deprecated. I am
not happy with any solution, but 2.3 release is near, so this wont
stay long.