[uggrid][bugfix] Don't add MPI flags explicitly
if you compile Dune with a sequential UG, but with MPI, the old configuration added MPI
to the CPPFLAGS. This implied that GridFactory< UGGrid<...> > use the MPI version of
CollectiveCommunication. When you now tried to run a seqential program, i.e. you didn't
add the MPI flags, you got an MPI error when creating a UGGrid using the GridFactory, as
you tried to call MPI command (in UG part of libdunegrid) without calling MPI_Init 
(as MPIHelper used the fake version). Instead of adding the MPI flags explicitly, we rely on the
buildsystem. The m4 test adds the MPI flags to teh UG flags, if a parallel UG and
MPI are found.