[release, buildsystem] Allow reuse of grid tests for 3rdparty grids with installed dune-grid.
The source of the grid test of dune-grid never got installed. Given their genericity they are
quite useful for downstream grid developers. These developers might want to be able to test
their grid even with an installed dune-grid module. To facilitate this the commited patch now
installs the test source.