[bugfix] Implemented {leaf,level}GridView for geometrygrid.
Previously code calling those methods broke due to the fallback
to the grid default implementation which expects a different
consructor of the grid view. This patch simply prevents the usage
of the default implementations by adding matching functions to
geometry grid that call the already implemented {leaf,level}View
methods.