Implement load balancing with prescribed element target ranks and data transfer
The previous patch allowed to load balance a grid according to an array of
prescribed target ranks, one for each elements.  This patch adds another method
which does the same thing, but additionally allows to transport data with it.
As in the case with automatic repartitioning, only node data is currently
supported.