[VTK] Print out an error message if the output file cannot be created
If the VTKWriter fails to create the output file (most of the time
because the containing directory does not exist), the program fails with
a very unhelpful error message.

This patch improves the situation by at least logging to stderr that
something went wrong during the file creation before re-throwing the
exception.