Add method Entity<0>::count (unsigned int codim) to the grid interface
Currently, the method exists, but the argument is a template argument
rather than a run-time parameter.  It was decided in

https://dune-project.org/flyspray/index.php?do=details&task_id=1455

to change this.  A previous patch has added this method to the grid
implementations.  This patch adds the method to the interface.
This does not break third-party grid implementations that do not
implement the method, as long as this method is not actually called.