[MCMGMapper][Bugfix] Fix assertion to check whether GeometryType is part of the mapper
The assertion "GlobalGeometryTypeIndex::index(gt) < n" doesn't make
sense, as n refers to the number of entities in the mapper and not to
the indexed GeometryTypes.

Fixed by simply reusing the method contains() conveniently provided by
the layout... ;-)

While I was at it, I also added an equivalent assertion to the plain
map() method that maps an entity itself.