Use std::bitset<dim> instead of Dune::FieldVector<ctype,dim> for the shift
The shift of a YGrid is used to differentiate element Ygrids from vertices
Ygrids etc. However, their values are only a combination of 0 and 0.5 .
The same behaviour can be achieved by using a bitset and hardcoding the shift 0.5
if an entry is true.
Additionally, the shift isnt really needed in the sizeArray helper function
and is thus removed.