[GridView] Add support for using range-based for loops on GridView
These changes make it possible to use the C++11 range-based for
loop to iterate over entities in a GridView given either
dimension or codimension:

 using Dune::Codim;
 using Dune::Dim;
 for (auto& face : entities(gridview, Codim<1>())) { ... }
 for (auto& vertex : entities(gridview, Dim<0>())) { ... }

For common cases a bit nicer syntax is also available:

 for (auto& e : elements(gridview)) { ... }
 for (auto& f : faces(gridview)) { ... }
 for (auto& e : edges(gridview)) { ... }
 for (auto& v : vertices(gridview)) { ... }

It is not necessary for users to qualify the functions with the
Dune namespace as they are found using argument-dependent lookup.

This patch requires accompanying changes in dune-common (the
IteratorRange class) and dune-geometry (Dune::Dim and Dune::Codim).