[Iterator] make Iterator a forward iterator
as decided on the 2014 meeting the Level- & LeafIterator should be
compatible with the forward iterator concept. This means in particular
that we have to add default constructor and post-increment operator.
This patch implements the post-increment in the facades class. The
default constructor is added to the facades class, but it clearly
requires changes to the imeplementations.