[Common] Add PartitionSet class that provides a static wrapper for sets of PartitionTypes
In order to provide entity ranges for different PartitionIteratorTypes,
we need a class that captures a set of partitions in a type signature.

The new class PartitionSet does exactly that and also allows for
combining and subtracting such sets. It is also possible to statically
extract the PartitionIteratorType value that corresponds to the stored
set of partitions.

Finally, it adds a set of default PartitionSets to the namespace
Dune::Partitions. Those instances are placed in an anonymous namespace
to avoid linking problems and cover all of the current PartitionType
values and their iterable combinations.