Start working on a backup restore facility in YaspGrid
With the tensorproduct feature it might be interesting for some
users to export the grid. As this should not only be for actually
restoring the grid, but also for postprocessing data, the output
format will be readable by the human eye.

As a starting point, we shamelessly copy alugrids facility.