Introduce a comm parameter on the restore methods
This change is kind of experimental, as it
a) implements an interface change under discussion
b) needs a default constructor on the CollectiveCommunication
c) does not work yet in sequential AND parallel

The latter is because of the inconsistencies in Yasps
handling of communicators throughout there constructors.
I will have to look into that.