Deprecated the file sgrid.hh
At the 2014 dev meeting in Berlin, it was decided that the SGrid grid manager
should be removed from dune-grid after the 2.4 release.  Rationale: YaspGrid
now implements all features of SGrid, and there is no reason to have the
maintenance burden of both grids.  This patch hence deprecates SGrid in the 2.4
release, and announces its removal.