[Documentation] Document range-based for helpers
This patch adds a whole bunch of documentation for the new helper
functions that enable iteration over entities and intersections using
range-based for.

There is a general introduction to the concept and detailed
documentation for each of the functions. The functions are also listed
as related to the GridView, so they also appear on that documentation
page.

The functions are added to a new subgroup of the grid interface group.